pyproject.toml
[tool.black]
line-length = 90
target-version = ['py310', 'py311', 'py312']
skip-string-normalization = false

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = 'TexSoup.*'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'latexcodec.*'
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests on the full acl-org/acl-anthology repo"
]
testpaths = ["tests"]

[tool.ruff]
ignore = ['E501']  # "Line too long" is black's job
target-version = 'py310'

[tool.poetry]
name = "acl-anthology-py"
packages = [
    { include = "acl_anthology" },
]
version = "0.4.2"
description = "A library for accessing the ACL Anthology"
authors = ["Marcel Bollmann <marcel@bollmann.me>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/mbollmann/acl-anthology-py"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11.0 || >3.11.0,<3.13"
docopt = "^0.6.2"
latexcodec = "^2.0.1"
lxml = "^4.9.2"
PyYAML = "^6.0"
app-paths = "^0.0.7"
diskcache = "^5.6.1"
rich = "^13.3.5"
attrs = "^23.1.0"
texsoup = ">=0.3.1,<0.4.0"
omegaconf = "^2.3.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.1"}
numpy = "^1.26.0"
scipy = "^1.6.0"
typing-extensions = {version = "^4.6.0", python = "<3.11"}
platformdirs = "^3.11.0"
gitpython = "^3.1.37"

[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
ruff = "^0.0.269"
pre-commit = "^3.3.2"
richbench = "^1.0.3"
mkdocs = "^1.4.3"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mkdocs-material = "^9.1.15"
mypy = "^1.3.0"
types-python-slugify = "^8.0.0.3"
types-pyyaml = "^6.0.12.11"
types-lxml = "^2023.3.28"
timer4 = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
